% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare Two Matrices by 'Vector Field Consensus'}
\usage{
compare(x, y)
}
\arguments{
\item{x}{matrix.}

\item{y}{matrix.}
}
\value{
numeric.
}
\description{
Compare two matrices to identify indices of point correspondence by 'Vector Field Consensus'
}
