% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask}
\alias{mask}
\title{Mask Two Images by 'Vector Field Consensus'}
\usage{
mask(x, y)
}
\arguments{
\item{x}{matrix.}

\item{y}{matrix.}
}
\value{
numeric.
}
\description{
Mask two images to identify indices of point correspondence by 'Vector Field Consensus'
}
