% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvfc-package.R
\docType{package}
\name{rvfc-package}
\alias{rvfc}
\alias{rvfc-package}
\title{rvfc: Experimental R bindings for 'Vector Field Consensus' with OpenCV}
\description{
Vector field consensus (VFC) is an efficient algorithm for establishing robust point correspondences between two sets of points. The C++ / OpenCV source code for the VFC algorithm can be downloaded from GitHub: \url{https://github.com/jiayi-ma/VFC/blob/master/VFC_opencv.zip}. Several elements of the opencv package configuration have been utilized in this package for loading the OpenCV libraries and facilitating compilation (c/o Jeroen Oooms)
}
\details{
For additional details on the algorithm, see: \url{https://www.cs.jhu.edu/~ayuille/Pubs10_12/TIP2014.pdf}
}
\references{
Jiayi Ma, Ji Zhao, Jinwen Tian, Alan Yuille, and Zhuowen Tu. (2014) "Robust Point Matching via Vector Field Consensus", IEEE Transactions on Image Processing, 23(4), pp. 1706-1721. doi: 10.1109/TIP.2014.2307478.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://humus.rocks/rvfc}
  \item Report bugs at \url{https://github.com/brownag/rvfc/issues}
}

}
\author{
\strong{Maintainer}: Andrew Brown \email{brown.andrewg@gmail.com}

Other contributors:
\itemize{
  \item Jeroen Ooms \email{jeroen@berkeley.edu} (\href{https://orcid.org/0000-0002-4035-0289}{ORCID}) (Rcpp configuration and cv::Mat-related glue code from {opencv}.) [contributor]
}

}
\keyword{internal}
